+++ app/src/main/java/app/black0ut/de/map_service_android/data/Strategy.java
@@ -19,7 +19,7 @@ public class Strategy {
     @Expose
     public Boolean [] drag;
     @Expose
-    public Integer [] x;
+    public Float [] x;
     @Expose
-    public Integer [] y;
+    public Float [] y;
 }
